changed authcomp.2 to authcomp. ( name )
