+++ admin/src/features/authByGoogle/ui/AuthComponent/AuthComponent.tsx
@@ -1,17 +1,19 @@
 import { app } from "../../../../../firebase"
 import { GoogleAuthProvider, getAuth, signInWithPopup, signOut } from "firebase/auth"
-import { useContext, useEffect } from "react"
+import { useContext, useEffect, useRef, useState } from "react"
 import { AuthContext } from "../../context/AuthContext"
 import styles from "./AuthComponent.module.scss"
 import { checkIsEmailWhitelisted } from "../../lib/checkIsEmailWhitelisted"
 import { Alert } from "shared/ui/CustomNotifications"
+import { DownArrowIcon, ExitIcon } from "shared/assets/icons"
+import { DELAY } from "../../const/delay"
 
 const auth = getAuth(app)
 const provider = new GoogleAuthProvider()
 
 export function AuthComponent() {
+    const [dropdownCheckbox, setDropdownCheckbox] = useState(false)
     const { user, updateUser } = useContext(AuthContext)
-
     function logOutWithGoogle() {
         updateUser(null)
         signOut(auth)
@@ -32,6 +34,30 @@ export function AuthComponent() {
                 Alert.error("Auth error")
                 logOutWithGoogle()
             })
+    }
+
+    const dropdownTimeout = useRef<ReturnType<typeof setTimeout> | null>(null)
+
+    function dropdownSwitcher() {
+        setDropdownCheckbox(prev => {
+            if (!prev) {
+                if (dropdownTimeout.current) {
+                    clearTimeout(dropdownTimeout.current)
+                }
+                dropdownTimeout.current = setTimeout(() => setDropdownCheckbox(false), DELAY)
+                return true
+            } else {
+                return false
+            }
+        })
+    }
+
+    function mouseEnterHandler() {
+        clearTimeout(dropdownTimeout.current as any)
+    }
+
+    function mouseLeaveHandler() {
+        dropdownTimeout.current = setTimeout(() => setDropdownCheckbox(false), DELAY)
     }
 
     useEffect(() => {
@@ -55,11 +81,38 @@ export function AuthComponent() {
     return (
         <>
             {user ? (
-                <div className={styles.logOut}>
-                    <div>
-                        {user.name} <img src={user.img} />
+                <div className={styles.container}>
+                    <div className={styles.sliderContainer}>
+                        <div>
+                            <label htmlFor="dropdowncheckbox" className={styles.slider}>
+                                <img src={user.img} alt="user image" className={styles.avatar} />
+                                <p className={styles.name}>{user.name}</p>
+                                <DownArrowIcon className={styles.downarrow} />
+                            </label>
+                            <input
+                                type="checkbox"
+                                id="dropdowncheckbox"
+                                checked={dropdownCheckbox}
+                                onChange={dropdownSwitcher}
+                                className={styles.dropdownCheckbox}
+                            />
+                        </div>
+                        <div
+                            className={`${styles.dropdown} ${
+                                dropdownCheckbox ? styles.dropdownVisible : styles.dropdownInvisible
+                            }`}
+                        >
+                            <button
+                                className={styles.logOutBtn}
+                                onClick={logOutWithGoogle}
+                                onMouseEnter={mouseEnterHandler}
+                                onMouseLeave={mouseLeaveHandler}
+                            >
+                                Log Out
+                                <ExitIcon className={styles.exitIcon} />
+                            </button>
+                        </div>
                     </div>
-                    <button onClick={logOutWithGoogle}>Log Out</button>
                 </div>
             ) : (
                 <button onClick={signInWithGoogle}>Sign in</button>
